Add libclinic.parse_args module
